refactor(profile): extract shared PersonCard + name-cache hook [quality-012]
Dedupe the copy-pasted PersonCard and module-scoped name caches across the endorsements/followers profile tabs into one shared component + hook.
